requirements.txt
# Core dependencies
pathspec>=0.11.0  # For gitignore-style pattern matching
PyYAML>=6.0  # For .pre-commit-config.yaml handling

# Development dependencies
pytest>=7.0.0
pre-commit>=3.0.0

# Optional: for code quality
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
